move model and dataset to the core API
